refactor: repository 인터페이스 domain 계층으로 이동
