+++ src/main/java/roomescape/domain/reservations/entity/ReservationTimeRepository.java
@@ -1,8 +1,7 @@
-package roomescape.domain.reservations.infrastructure;
+package roomescape.domain.reservations.entity;
 
 import java.util.List;
 import java.util.Optional;
-import roomescape.domain.reservations.entity.ReservationTime;
 
 public interface ReservationTimeRepository {
     ReservationTime save(ReservationTime reservation);
