+++ src/main/java/roomescape/domain/reservations/infrastructure/ReservationJdbcTemplateRepository.java
@@ -8,6 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.reservations.entity.Reservation;
+import roomescape.domain.reservations.entity.ReservationRepository;
 import roomescape.domain.reservations.entity.ReservationTime;
 
 @Repository
