+++ src/test/java/roomescape/domain/reservations/FakeReservationTimeRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import roomescape.domain.reservations.entity.ReservationTime;
-import roomescape.domain.reservations.infrastructure.ReservationTimeRepository;
+import roomescape.domain.reservations.entity.ReservationTimeRepository;
 
 public class FakeReservationTimeRepository implements ReservationTimeRepository {
